fix: Resolve build errors and add multi-core synchronization
Fixes #8

## Build Fixes
- Added Tools/compat/elf.h for macOS ELF definitions
- Replaced Linux-specific memfd_create/copy_file_range with ANSI C
- Removed unsupported -fropi/-frwpi flags and -lbsd dependency
- Fixed missing MaskedInterruptGuard.hpp includes
- Ensured 8-byte stack alignment to prevent hardfaults

## New Features
- SoftwareSpinLock with deadlock detection and wfe/sev
- SoftwareMutex for blocking synchronization
- WaitingThreadQueue for thread management
- Power-efficient multi-core sync primitives

## Bug Fixes
- Fixed HardwareSpinLock infinite loop on lock acquisition
- Fixed memory leaks in VirtualFileHandle
- Fixed stat /dev/tty returning invalid device info
- Fixed ConsoleFileHandle invalid pointer issues

## Documentation
- Restructured Docs/ with architecture guides
- Added Mermaid diagrams for visualization
- Created syscall reference and shell command docs

## Configuration
- Added VSCode IntelliSense and Cortex-Debug support
- Updated CMakeLists.txt with version info
- Reformatted CHANGELOG.md following Keep a Changelog

+++ Tests/Std/TestMemoryAllocator.cpp
@@ -53,7 +53,7 @@ TEST_CASE(memoryallocator_death_by_a_thousand_cuts)
     ASSERT(stats_middle.m_largest_continous_block < heap.size() / 4);
 
     // We want to deallocate in a random, but reproducible order
-    std::mt19937 prng { 3040088493306752707 };
+    std::mt19937 prng { static_cast<unsigned int>(3040088493306752707) };
     std::shuffle(allocations.begin(), allocations.end(), prng);
 
     for (u8 *allocation : allocations)
